Move AMQPProxy and ConsumerPool into their own modules
This is a minor change in how the modules work and should not affect any external interfaces
