+++ amqpdispatcher/dispatcher.py
@@ -8,8 +8,8 @@ import socket
 import sys
 import urlparse
 
+from amqpdispatcher.consumer_pool import ConsumerPool
 from haigha.connections.rabbit_connection import RabbitConnection
-from haigha.message import Message
 from yaml import safe_load as load
 import gevent
 import gevent.queue
@@ -211,91 +211,6 @@ def load_module_object(module_object_str):
     module_name, obj_name = module_object_str.split(':')
     module = load_module(module_name)
     return getattr(module, obj_name)
-
-
-class ConsumerPool(object):
-    def __init__(self, channel, klass, greenlet_maker, size=1):
-        self._channel = channel
-        self._pool = gevent.queue.Queue()
-        self._klass = klass
-        self._gm = greenlet_maker
-        for i in range(size):
-            self._create()
-
-    def _create(self):
-        logger.debug('Creating consumer instance: {0}'.format(
-            self._klass.__name__
-        ))
-        self._pool.put(self._klass())
-
-    def handle(self, msg):
-        def func():
-            consumer = self._pool.get()
-            amqp_proxy = AMQPProxy(self._channel, msg)
-
-            def put_back(successful_greenlet):
-                logger.debug('Successful run, putting consumer back')
-                if not amqp_proxy.has_responded_to_message:
-                    amqp_proxy.ack()
-                self._pool.put(consumer)
-
-            def recreate(failed_greenlet):
-                logger.info('Consume failed, shutting down consumer')
-                if not amqp_proxy.has_responded_to_message:
-                    amqp_proxy.reject(requeue=True)
-                shutdown_greenlet = gevent.Greenlet(
-                    consumer.shutdown,
-                    failed_greenlet.exception
-                )
-
-                def create_wrapper(*args):
-                    self._create()
-                shutdown_greenlet.link(create_wrapper)
-                shutdown_greenlet.start()
-
-            greenlet = self._gm(consumer.consume, amqp_proxy, msg)
-            greenlet.link_value(put_back)
-            greenlet.link_exception(recreate)
-            greenlet.start()
-        self._gm(func).start()
-
-
-class AMQPProxy(object):
-
-    def __init__(self, channel, msg):
-        self._channel = channel
-        self._msg = msg
-        self._terminal_state = False
-
-    @property
-    def tag(self):
-        return self._msg.delivery_info['delivery_tag']
-
-    @property
-    def has_responded_to_message(self):
-        return self._terminal_state
-
-    def ack(self):
-        self._error_if_already_terminated()
-        self._channel.basic.ack(self.tag)
-
-    def nack(self):
-        self._error_if_already_terminated()
-        self._channel.basic.nack(self.tag)
-
-    def reject(self, requeue=True):
-        self._error_if_already_terminated()
-        self._channel.basic.reject(self.tag, requeue=requeue)
-
-    def publish(self, exchange, routing_key, headers, body):
-        msg = Message(body, headers)
-        self._channel.basic.publish(msg, exchange, routing_key)
-
-    def _error_if_already_terminated(self):
-        if self._terminal_state:
-            raise Exception('Already responded to message!')
-        else:
-            self._terminal_state = True
 
 
 def message_pump_greenthread(connection):
