+++ tests/unit/test_amqp_proxy.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 
 from mock import MagicMock
 
-from amqpdispatcher.dispatcher import AMQPProxy
+from amqpdispatcher.amqp_proxy import AMQPProxy
 
 
 class TestAmqpProxy(TestCase):
